display detections results
