+++ app/src/main/java/com/example/visionapp/ui/screens/HomeScreen.kt
@@ -2,7 +2,6 @@ package com.example.visionapp.ui.screens
 
 import android.Manifest
 import android.graphics.Bitmap
-import android.util.Log
 import android.widget.Toast
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
@@ -21,7 +20,6 @@ import com.example.visionapp.CameraConfig
 import androidx.camera.lifecycle.ProcessCameraProvider
 import com.example.visionapp.ModelsConfig
 import com.example.visionapp.communiates.CommunicateGenerator
-import com.example.visionapp.communiates.CommunicateQueue
 import com.example.visionapp.model.TriangleMethod
 import com.example.visionapp.onnxmodels.ModelPredictor
 import com.example.visionapp.onnxmodels.models.DepthModel
@@ -33,7 +31,8 @@ import com.example.visionapp.onnxmodels.processing.DetectionResult
 import com.example.visionapp.onnxmodels.processing.SegmentationPostprocessor
 import com.example.visionapp.utils.startCameraWithAnalyzer
 import com.example.visionapp.onnxmodels.ModelType
-import com.example.visionapp.onnxmodels.processing.SegmentationBitmapHelper
+import com.example.visionapp.onnxmodels.processing.Helpers.DetectionBitmapHelper
+import com.example.visionapp.onnxmodels.processing.Helpers.SegmentationBitmapHelper
 import com.example.visionapp.processing.DetectionProcessing
 import com.example.visionapp.ui.common.DropdownMenuControl
 import com.example.visionapp.utils.TextToSpeechManager
@@ -108,7 +107,11 @@ fun HomeScreen(navController: NavController) {
         if(ModelsConfig.VISUAL_DEBUG_MODE){
             when (selectedModelDebug){
                 ModelType.DEPTH -> depthImage?.let { addBitmapToBuffer(it) }
-                ModelType.DETECTION -> {} //TODO
+                ModelType.DETECTION -> {
+                    val scaledImageBitmap = scaleBitmap(bitmap, CameraConfig.DETECTION_RESOLUTION)
+                    val imageWithBoxes = DetectionBitmapHelper.drawDetectionsOnBitmap(scaledImageBitmap, detectionResults)
+                    addBitmapToBuffer(imageWithBoxes)
+                }
                 ModelType.SEGMENTATION -> segmentedImage?.let{
                     val scaledImageBitmap = scaleBitmap(bitmap, CameraConfig.SEGMENTATION_RESOLUTION)
                     val imageWithOverlay = SegmentationBitmapHelper.overlayColoredMaskOnImage(segmentedImage, scaledImageBitmap)
